sendMail function and email contact added. Unified contact controller
