+++ routes/contacts.js
@@ -1,7 +1,6 @@
 
 const express = require("express");
-const { createContact, contactFindAll } = require("../controllers/contact.controller");
-const { getContacts } = require('../controllers/contacts');
+const { getContacts, createContact, contactFindAll } = require('../controllers/contacts');
 const { checkToken} = require('../middlewares/check-token');
 const {checkIsAdminLogged} = require('../middlewares/commons');
 const router = express.Router();
